pyproject.toml
[project]
name = "cloudrun-scheduler-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.27.0",
]
